Make "testing" provider auto-installable for the test config
This provider is a third-party community provider (developed by me) which
previously needed to be manually installed, because at the time we
published this module Terraform could only auto-install HashiCorp or
HashiCorp-partner providers.

Terraform 0.13 introduces an open ecosystem for community providers and
this "testing" provider is now available via that, so we'll declare that
here and thus make this module more self-contained.

The testing module remains experimental, so the version pin here is to
exactly 0.0.1 in case subsequent versions include breaking API changes.